use Gather/Scatter instead of distributive laws
Fixes #50.

Distributive monads and comonads looked like a principled way to
combine recursion schemes, by reusing the way in which monad and
comonad transformers combine. For this approach to work, we also need
to combine the distributive laws, but unfortunately this is error
prone because sometimes the only implementation which type-checks,
e.g. `distZygoT`, produces distributive laws which aren't lawful, e.g.
`distZygoT g distHisto`. See #50 for details.

Here is an alternative approach.

For folds, the idea is that the information should always flow from
the leaves to the root. For `cata`, the f-algebra computes an `a` for
the leaves, and then uses those `a`s to compute an `a` for their
parents, and then for their grand-parents, and so on. For `gcata`, the
algebra still computes an `a` for the leaves, but when computing the
`a` for a node from the information computed for the direct children,
that information includes more than just an `a`; it could be `(t,a)`,
or `Cofree f a`, etc. -- let's call it `s`. Since the algebra only
computes an `a` from this `f s`, something else, a "gathering
function", needs to compute the rest of the `s` from this `f s`. We
can combine those gathering functions in the same way we were
combining the distributive laws; except this time `gatherZygoT g
gatherHisto` behaves sensibly.

For unfolds, the dual idea is that the information flows from the root
to the leaves. For `ana`, the f-coalgebra computes an `f a` from the
root's `a`, thereby creating the root's direct children, and then we
use those `a`s to compute the `a`s for the grand-children, and so on.
For `gcata`, the coalgebra still computes an `f` from an `a`, but it
has the opportunity to describe the sub-trees using other
representation than an `a`; it could be `Either t a`, or `Free f a`,
etc. -- let's call it `s`. Since the coalgebra only knows how to
expand an `a` into an `f s`, something else, a "scattering function",
needs to handle the rest of `s`'s constructors.