Documentation "improvements"
It turns out the README was doubly-wrong. The README claims that upon
seeing the word "para", recursion-schemes users immediately understand
that they are looking at a "spine-based" function, which updates the
nodes along a path from the root to a leaf while leaving the rest of the
nodes untouched. Unfortunately:

1. "spine-based" is a terrible way of describing this, because the spine
   includes all the internal nodes, not just those along the path from
   the root to that particular leaf. A more established word for
   describing this kind of algorithm is "path-copying".
2. recursion-schemes users do _not_ immediately understand that they are
   looking at a path-copying function when they see the word "para", for
   the good reason that only a small fraction of the uses of para are
   path-copying functions. Indeed, the haddocks for para showcases an
   example which is not a path-copying function!

The first mistake is easy to fix, but unfortunately the second is not:
being able to get that kind of big picture information at a glance
sounds like such a great example of the benefits of recursion-schemes
that removing it makes the rest of the README fall short on its attempt
to sell the library.

To recap: the README was previously wrong (but I didn't know it was),
then the para example was added and now the documentation is both wrong
and self-contradictory. This PR thus deletes the (correct!) example in
order to revert back to the slightly-less-bad state in which the
documentation is merely wrong, not self-contradictory.

Of course, in the future, I would like to take the time to make the
documentation correct, either by adding a new, specialized version of
para to the library which _does_ convey that big picture information, or
by finding another example which sounds as good, or by rewriting the
README to list fewer benefits. We can restore the example then.