RecursionSchemes: more readable and more correct
This is a huge, yet backwards-compatible change! It introduces a new API
which is both more readable and more correct than the previous API, and
it also changes the name of the recursion-schemes import.

First, the name of the module. "Data.Functor.Foldable" is a strange
choice of module name by modern standards, "RecursionSchemes" is better
because it matches the name of the library instead of trying to fit into
the "Data.<...>" / "Control.<...>" semantic hierarchy. By introducing
the new API and the new module name at the same time, it becomes
possible to preserve backwards compatibility simply by keeping the old
API at the old module name and adding the new API at the new module
name.

Second, the new API. One big advantage of Mendler-style
recursion-schemes such as `mcata` and `mhisto` is that the resulting
code is more readable. At least in my opinion, since readability is
always subjective. Still, compare the following four implementations of
`fib`:

    data Unary = Zero | Succ Unary
    makeBaseFunctor ''Unary

    fibA :: Unary -> Int
    fibA = \case
      Zero
        -> 1
      Succ Zero
        -> 1
      Succ nMinus1@(Succ nMinus2)
        -> fibA nMinus1 + fibA nMinus2

    fibB :: Unary -> Int
    fibB = histo $ \case
      ZeroF
        -> 1
      SuccF (_ :< ZeroF)
        -> 1
      SuccF (fibNMinus1 :< SuccF (fibNMinus2 :< _))
        -> fibNMinus1 + fibNMinus2

    fibC :: Fix (Base Unary) -> Int
    fibC = mhisto $ \recur split -> \case
      ZeroF
        -> 1
      SuccF nMinus1 -> case split nMinus1 of
        ZeroF
          -> 1
        SuccF nMinus2
          -> recur nMinus1 + recur nMinus2

    fibD :: RecFun Unary Int
    fibD = recFun histo $ \case
      ZeroF
        -> 1
      SuccF nMinus1 -> case project nMinus1 of
        ZeroF
          -> 1
        SuccF nMinus2
          -> recur nMinus1 + recur nMinus2

I would say that `fibA` is the most readable, as it directly expresses
that `fib 0 = 1`, `fib 1 = 1`, and `fib n = fib (n-1) + fib (n-2)`. But
that definition uses general recursion, which does not guarantee that
we're only making recursive calls on smaller terms. It is also very
inefficient.

`fibB` improves upon `fibA` by using the `histo` recursion scheme, which
perform the recursion on `fibB`'s behalf; and since `fibB` does not
make any recursion calls itself, it cannot accidentally make recursive
calls on terms which are not smaller. It is also much more efficient,
because the recursion structure is no longer branching into two, instead
it combines two pre-computed results. However, `fibB`'s pattern-matching
is a lot messier: the structure of the Unary and the pre-computed
results are interleaved, and you have to be very familiar with the
details of the `Base Unary (Cofree (Base Unary) Int)` type on which it
is pattern-matching in order to know which position holds what. This
becomes even messier when combining multiple recursion-schemes, e.g.
with a zygo-paramorphism.

Furthermore, while the word "histo" is helpful to those who are familiar
with that name, as it immediately conveys the idea that we are going to
be combining recursively-pre-computed values from multiple depths, it is
deeply unhelpful to those who aren't, as the recursion structure of
`fibB` is no longer explicit, it is now hidden inside the implementation
of `histo`.

Next, `fibC` has the best of both worlds: the word `mhisto` conveys
the same idea to those who are familiar with the name "histo", while the
explicit `recur` calls clarify the recursion structure for those who
aren't. Furthermore, while the pattern-matching is not as readable as in
`fibA`, it is much more readable than in `fibB`, because we don't have
to bring all of the information into scope all at once. Instead, the
helper functions `recur` and `split` extract the bits we need when we
need them. We also get the same efficient implementation; `recur`
extracts the pre-computed value for that recursive position, it does not
actually make a recursive call, but `recur` is still a very good name
for it because it returns the same result as if we were making a
recursive call to `fibC`. Moreover, the type of `recur` is restricted in
a way which guarantees that it can only be called on smaller terms.

Finally, `fibD` demonstrates the new API. Like `fibC`, it has the best
of both worlds, but in addition, the `recur` and `split` (renamed to
`project`, because we can reuse `Recursive.project`) functions are now
global instead of local. Also, while `Data.Functor.Foldable.histo` has a
`ghisto` variant which makes it possible to combine recursion-schemes,
`Data.Functor.Foldable.mhisto` doesn't. In contrast, the new API's
design does make it possible to provide a `histoT` variant while
retaining the advantages of the Mendler style.

See
https://github.com/ekmett/recursion-schemes/pull/105#issuecomment-694308336
for a similar example with `para` instead of `histo`.

One last advantage of the new API is that is is more correct than the
old API. See
https://github.com/ekmett/recursion-schemes/issues/50#issuecomment-396057219
for an explanation of what is wrong with the old API, and
https://github.com/ekmett/recursion-schemes/pull/51
for an explanation of the Gather/Scatter solution to that problem. In
short, generalized recursion-schemes such as `ghisto` can sometimes
combine in ways that appear type-correct but actually give
subtly-incorrect results, while the new API's transformers, such as
`histoT`, use a much simpler approach, in the hope of making the library
closer to "so simple that there are obviously no deficiencies" than to
"so complicated that there are no obvious deficiencies".

In the new API, everything is a catamorphism applied to an algebra of
type `base pos -> pos`. This `pos` is different for every recursion
scheme, and it contains all the data which the recursion scheme needs to
keep track during the catamorphism. The user provides a function of type
`base pos -> a`, so `pos` must also have instances which allow the user
to call functions like `recur` on the recursive positions. Each
recursion scheme picks their `pos` so that the user is only allowed to
make the observations which make sense for that recursion scheme. That's
it. Well, sort of.

The user provides a function of type `base pos -> a`, but we need a
function of type `base pos -> pos`. The user's function explains how one
step of the recursion computes the value the user is interested in from
the values provided by the `base pos`. What we need is a function
explaining how one step of the recursion computes all the data we need
in order to continue the fold, including both the data which the user is
interested in and the data which the next incarnation of the user
function will be able to draw from. Each recursion scheme must thus
provide a "gather" function of type `a -> base pos -> pos`, which
extracts all that information, and stores it in the `pos` along with the
`a` which the user function returned; after all, the next incarnation of
the user function must be able to call `recur` on that `pos` in order to
get back that `a`.

We might also want to combine recursion schemes; that's the feature for
which the old API brought in complex things like comonad transformers
and functions expressing distributive laws. In the new API, a recursion
scheme transformer such as `paraT` is associated with a `pos`
transformer, such as `ParaPosT`, and instead of providing a gather
function, it transforms a gather function on `pos` into a gather
function on `ParaPosT pos`.

The `Scatter` version, where everything is an anamorphism, will be
implemented in a latter commit.