hylo gets its own module
it's the black sheep of recursion schemes. it's a faster version of
(ana f >> cata g), but  (ana f >>> cata g) is a bad idea because it
doesn't guarantee termination: cata is guaranteed to terminate given a
finite input, but ana is only guaranteed to be productive, it does not
guarantee to generate a finite output.