re-generalize the type of depth and size
they type was originally `Tree a -> Int`. we wanted to demonstrate a
common mistake, so we changed the type to `Tree Int -> Int` so the
mistaken code would still typecheck. However, the mistaken code does not
actually typecheck. So there is no benefit to using the more specialized
type signature.

this also means that the example is not that great at demonstrating how
para is safer, because the mistake would have been caught by the
type-checker anyway. but since the mistake is very important for the
flow of the story, I think I'll leave it in.