control whether to generate Base, {Co,}Recursive
'makeBaseFunctor' is conveniently generates all of Base, Recursive, and
Corecursive. But what if you only want to generate some of those, e.g.
because you need to write a non-standard instance by hand, or because
your type should only be folded but never unfolded?

'makeBaseFunctorWith' now takes extra options to specify which of the
three should be generated. The new 'makeRecursive' and 'makeCorecursive'
are shortcuts for the common case in which 'Base' and only one of the
two instances are needed.