Improve inference for makeBaseFunctor
Previously, `MakeBaseFunctor` had both `[a]` and `Q a` instances,
which forced users to provide a type annotation for the declarations
style; this was awful, and not properly documented. Be more clever,
using overlapping instances to fix the type constructor. This still
allows `makeBaseFunctor` to be used with lists of names, but
otherwise infers everything nicely.