Support building with th-abstraction-0.5.*
`th-abstraction-0.5.0.0` introduces a `TypeData` constructor for
`DatatypeVariant` that characterizes `type data` declarations (introduced in
GHC 9.6). Because `type data` data constructors only exist at the type level,
not the value level, it doesn't make sense to define `Recursive` or
`Corecursive` instances for a `type data` declaration. As a result, this patch
makes it an error to use `recursion-schemes`' TH machinery in conjunction with
a `type data` declaration.

Fixes #142.