Drop support for pre-8.0 versions of GHC
This drops support for GHC 7.10 and earlier. This allows us to remove a
considerable amount of CPP. Moreover, now that we can use the
`TemplateHaskellQuotes` language extension without CPP, we use
`TemplateHaskellQuotes` to clean up the (rather gnarly) internals of
`Data.Functor.Foldable.TH`. Doing so makes it straightforward to fix #160, as
we no longer need to wire in the module names of `Functor`, `Foldable`,
`Traversable`, etc. directly into `Data.Functor.Foldable.TH`.

Fixes #160.